+++ view/base/requirejs-config.js
@@ -1,0 +1,12 @@
+var config = {
+    config: {
+        mixins: {
+            'Magento_Ui/js/form/element/image-uploader': {
+                'Overdose_Core/js/form/element/image-uploader-mixin': true
+            },
+            'Magento_Ui/js/form/element/file-uploader': {
+                'Overdose_Core/js/form/element/file-uploader-mixin': true
+            }
+        }
+    }
+};